Convert SkewableClock to a StoppableClock Skewable clock has concerns that it might lie to the underlying registries. This has been converted to a StoppableClock with the ability to stop the clock only exposed to the StepMeterRegistry.java.
